refactor(ramps): share resolveSdpEnvironment/rampRuntime via payments context
The two helpers were duplicated in handlers/ramps.ts and
handlers/ramps/bvnk.ts; both now import from routes/payments/context.
